refactor: verification methods dont need to live on the wallet
