+++ test/wallet.test.ts
@@ -1,4 +1,8 @@
-import Wallet, { formatDirectSignDoc } from '../src';
+import Wallet, {
+  formatDirectSignDoc,
+  verifyAminoSignature,
+  verifyDirectSignature,
+} from '../src';
 import {
   TEST_COSMOS_ADDRESS,
   TEST_COSMOS_AMINO_SIGNATURE,
@@ -41,11 +45,7 @@ describe('Wallet', () => {
     );
     const result = await wallet.signDirect(signerAddress, signDoc);
     expect(
-      wallet.verifyDirectSignature(
-        signerAddress,
-        result.signature.signature,
-        signDoc
-      )
+      verifyDirectSignature(signerAddress, result.signature.signature, signDoc)
     ).toBeTruthy();
     expect(result).toBeTruthy();
     expect(result.signature.signature).toEqual(TEST_COSMOS_DIRECT_SIGNATURE);
@@ -55,11 +55,7 @@ describe('Wallet', () => {
     const signDoc = TEST_COSMOS_INPUTS.amino;
     const result = await wallet.signAmino(signerAddress, signDoc);
     expect(
-      wallet.verifyAminoSignature(
-        signerAddress,
-        result.signature.signature,
-        signDoc
-      )
+      verifyAminoSignature(signerAddress, result.signature.signature, signDoc)
     ).toBeTruthy();
     expect(result).toBeTruthy();
     expect(result.signature.signature).toEqual(TEST_COSMOS_AMINO_SIGNATURE);
